requirements.txt
agency-swarm==0.2.6
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
attrs==24.2.0
black==24.8.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datamodel-code-generator==0.25.8
deepdiff==6.7.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
genson==1.3.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflect==5.6.2
instructor==1.3.4
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
jsonref==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.35.14
ordered-set==4.1.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
termcolor==2.4.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.9.4
ijson==3.1.4